remove excess stack/SV refcnting operations
No need to attach the SV * to mortal stack, its refcount is already held by
$self/HV *. PP delete() in non-void context uses sv_2mortal to delete the
SV * in the HE *, so it is unlikely a statement with many ","s and a delete
in an attempt to crash the code will succeed. Anyways, pp_hslice puts the
SV * in the HE * on Perl stack without a sv_2mortal.